Revert "Removed support for querying"
This reverts commit bbacd3316618cc333a431d8ae30de3950b508456.
