+++ test/test.ino
@@ -190,7 +190,7 @@ void testBasicFunction() {
     }
     TEST_ASSERT(client.isBufferEmpty());
     String query = "select";
-    String q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    String q = client.query(query);
     TEST_ASSERT(countLines(q) == 6);  //5 points+header
 
     // test precision
@@ -225,10 +225,14 @@ void testInit() {
     {
         InfluxDBClient client;
         String query = "select";
+        TEST_ASSERT(client.query(query) == "");
+        TEST_ASSERT(client.getLastStatusCode() == 0);
+        TEST_ASSERT(client.getLastErrorMessage() == "Unconfigured instance");
+
         client.setConnectionParams(INFLUXDB_CLIENT_TESTING_URL, INFLUXDB_CLIENT_TESTING_ORG, INFLUXDB_CLIENT_TESTING_BUC, INFLUXDB_CLIENT_TESTING_TOK);
         String rec = "a,a=1 a=3";
         TEST_ASSERT(client.writeRecord(rec));
-        String q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+        String q = client.query(query);
         TEST_ASSERT(countLines(q) == 2);  //3 points+header
     }
 
@@ -287,7 +291,7 @@ void testRetryOnFailedConnection() {
     delete p;
     TEST_ASSERT(clientOk.isBufferEmpty());
     String query = "select";
-    String q = queryFlux(clientOk.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    String q = clientOk.query(query);
     TEST_ASSERT(countLines(q) == 4);  //3 points+header
 
     TEST_END();
@@ -318,7 +322,7 @@ void testBufferOverwriteBatchsize1() {
     TEST_ASSERT(client.isBufferEmpty());
 
     String query = "select";
-    String q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    String q = client.query(query);
     int count;
     String *lines = getLines(q, count);
     TEST_ASSERTM(count == 6, String("6 != " + count));  //5 points+header
@@ -360,7 +364,7 @@ void testBufferOverwriteBatchsize5() {
     TEST_ASSERT(client.flushBuffer());
 
     String query = "select";
-    String q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    String q = client.query(query);
     int count;
     String *lines = getLines(q, count);
     TEST_ASSERT(count == 13);  //12 points+header
@@ -380,14 +384,14 @@ void testBufferOverwriteBatchsize5() {
         delete p;
     }
     TEST_ASSERT(!client.isBufferEmpty());
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     TEST_ASSERT(countLines(q) == 0);
 
     p = createPoint("test1");
     p->addField("index", 4);
     TEST_ASSERT(client.writePoint(*p));
     TEST_ASSERT(client.isBufferEmpty());
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     lines = getLines(q, count);
     TEST_ASSERT(count == 6);  //5 points+header
     TEST_ASSERT(lines[1].indexOf(",0") > 0);
@@ -416,7 +420,7 @@ void testServerTempDownBatchsize5() {
     }
     TEST_ASSERT(client.isBufferEmpty());
     String query = "select";
-    String q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    String q = client.query(query);
     TEST_ASSERT(countLines(q) == 16);  //15 points+header
     deleteAll(INFLUXDB_CLIENT_TESTING_URL);
 
@@ -439,7 +443,7 @@ void testServerTempDownBatchsize5() {
     p->addField("index", 15);
     TEST_ASSERT(client.writePoint(*p));
     TEST_ASSERT(client.isBufferEmpty());
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     TEST_ASSERT(countLines(q) == 17);  //16 points+header
     deleteAll(INFLUXDB_CLIENT_TESTING_URL);
 
@@ -460,7 +464,7 @@ void testServerTempDownBatchsize5() {
     waitServer(client, true);
 
     TEST_ASSERT(client.flushBuffer());
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     int count;
     String *lines = getLines(q, count);
     TEST_ASSERT(count == 21);  //20 points+header
@@ -491,7 +495,7 @@ void testRetriesOnServerOverload() {
     }
     TEST_ASSERT(client.isBufferEmpty());
     String query = "select";
-    String q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    String q = client.query(query);
     TEST_ASSERT(countLines(q) == 61);  //60 points+header
     deleteAll(INFLUXDB_CLIENT_TESTING_URL);
 
@@ -522,7 +526,7 @@ void testRetriesOnServerOverload() {
     TEST_ASSERT(!client.isBufferEmpty());
     TEST_ASSERT(client.flushBuffer());
     TEST_ASSERT(client.isBufferEmpty());
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     int count;
     String *lines = getLines(q, count);
     TEST_ASSERT(count == 40);  //39 points+header
@@ -558,7 +562,7 @@ void testRetriesOnServerOverload() {
     TEST_ASSERT(!client.isBufferEmpty());
     TEST_ASSERT(client.flushBuffer());
     TEST_ASSERT(client.isBufferEmpty());
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     lines = getLines(q, count);
     TEST_ASSERT(count == 40);  //39 points+header
     TEST_ASSERT(lines[1].indexOf(",11") > 0);
@@ -593,7 +597,7 @@ void testRetriesOnServerOverload() {
     TEST_ASSERT(!client.isBufferEmpty());
     TEST_ASSERT(client.flushBuffer());
     TEST_ASSERT(client.isBufferEmpty());
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     TEST_ASSERT(countLines(q) == 51);  //50 points+header
     lines = getLines(q, count);
     TEST_ASSERT(count == 51);  //50 points+header
@@ -630,7 +634,7 @@ void testRetriesOnServerOverload() {
     TEST_ASSERT(client.flushBuffer());
     TEST_ASSERT(client.isBufferEmpty());
 
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     lines = getLines(q, count);
     TEST_ASSERT(count == 40);  //39 points+header
     TEST_ASSERT(lines[1].indexOf(",11") > 0);
@@ -659,7 +663,7 @@ void testFailedWrites() {
     }
     int count;
     String query = "";
-    String q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    String q = client.query(query);
     String *lines = getLines(q, count);
     TEST_ASSERT(count == 17);  //16 points+header
     TEST_ASSERTM(lines[1].indexOf(",1") > 0, lines[1]);
@@ -682,7 +686,7 @@ void testFailedWrites() {
         delete p;
     }
 
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     lines = getLines(q, count);
     //3 batches should be skipped
     TEST_ASSERT(count == 16);  //15 points+header
@@ -725,7 +729,7 @@ void testTimestamp() {
     }
     int count;
     String query = "";
-    String q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    String q = client.query(query);
     String *lines = getLines(q, count);
     TEST_ASSERT(count == 21);  //20 points+header
     for (int i = 1; i < count; i++) {
@@ -747,7 +751,7 @@ void testTimestamp() {
         TEST_ASSERTM(client.writePoint(*p), String("i=") + i);
         delete p;
     }
-    q = queryFlux(client.getServerUrl(),INFLUXDB_CLIENT_TESTING_TOK, INFLUXDB_CLIENT_TESTING_ORG, query);
+    q = client.query(query);
     lines = getLines(q, count);
     TEST_ASSERT(count == 21);  //20 points+header
     for (int i = 1; i < count; i++) {
